+++ packages/dropdown_button2/example/custom_dropdown_button2.dart
@@ -78,7 +78,7 @@ class CustomDropdownButton2 extends StatelessWidget {
         ),
         value: value,
         items: dropdownItems
-            .map((String item) => DropdownMenuItem<String>(
+            .map((String item) => DropdownItem<String>(
                   value: item,
                   child: Container(
                     alignment: valueAlignment,