+++ packages/dropdown_button2/lib/dropdown_button2.dart
@@ -1,3 +1,4 @@
 library dropdown_button2;
 
 export 'src/dropdown_button2.dart';
+export 'src/dropdown_item.dart';