+++ packages/dropdown_button2_test/lib/src/few_styling_example.dart
@@ -52,7 +52,7 @@ class _FewStylingExampleState extends State<FewStylingExample> {
               ],
             ),
             items: items
-                .map((String item) => DropdownMenuItem<String>(
+                .map((String item) => DropdownItem<String>(
                       value: item,
                       child: Text(
                         item,