+++ packages/dropdown_button2_test/lib/src/popup_image_example.dart
@@ -31,14 +31,14 @@ class _PopupImageExampleState extends State<PopupImageExample> {
             openWithLongPress: true,
             items: [
               ..._MenuItems.firstItems.map(
-                (item) => DropdownMenuItem<_MenuItem>(
+                (item) => DropdownItem<_MenuItem>(
                   value: item,
                   child: _MenuItems.buildItem(item),
                 ),
               ),
-              const DropdownMenuItem<Divider>(enabled: false, child: Divider()),
+              const DropdownItem<Divider>(enabled: false, child: Divider()),
               ..._MenuItems.secondItems.map(
-                (item) => DropdownMenuItem<_MenuItem>(
+                (item) => DropdownItem<_MenuItem>(
                   value: item,
                   child: _MenuItems.buildItem(item),
                 ),