+++ packages/dropdown_button2_test/lib/src/simple_example.dart
@@ -32,7 +32,7 @@ class _SimpleExampleState extends State<SimpleExample> {
               ),
             ),
             items: items
-                .map((String item) => DropdownMenuItem<String>(
+                .map((String item) => DropdownItem<String>(
                       value: item,
                       child: Text(
                         item,